--- app/src/main/java/edu/agh/susgame/front/ui/component/common/Header.kt
@@ -1,24 +1,0 @@
-package edu.agh.susgame.front.ui.component.common
-
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.em
-import edu.agh.susgame.front.ui.theme.PaddingL
-
-@Composable
-fun Header(title: String) {
-    Text(
-        text = title,
-        fontSize = 6.em,
-        fontWeight = FontWeight.Bold,
-        textAlign = TextAlign.Center,
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(bottom = PaddingL)
-    )
-}
